Quarantine partial files with duplicates
