+++ app/file_sender.py
@@ -1,6 +1,4 @@
 import csv
-import hashlib
-import json
 import logging
 from datetime import datetime
 from pathlib import Path
@@ -11,7 +9,8 @@ from structlog import wrap_logger
 
 import app.sftp as sftp
 from app.encryption import pgp_encrypt_message
-from app.mappings import PRODUCTPACK_CODE_TO_DESCRIPTION, PACK_CODE_TO_DATASET, \
+from app.manifest_file_builder import generate_manifest_file
+from app.mappings import PACK_CODE_TO_DATASET, \
     SUPPLIER_TO_SFTP_DIRECTORY, DATASET_TO_SUPPLIER, DATASET_TO_PRINT_TEMPLATE
 from config import Config
 
@@ -53,8 +52,10 @@ def delete_local_files(file_paths: Iterable[Path]):
         file_path.unlink()
 
 
-def quarantine_partial_file(print_file: Path):
-    pass
+def quarantine_partial_file(partial_file_path: Path):
+    quarantine_destination = Config.QUARANTINED_FILES_DIRECTORY.joinpath(partial_file_path.name)
+    partial_file_path.replace(quarantine_destination)
+    logger.info('Quarantined partial print file', quarantined_file_path=str(quarantine_destination))
 
 
 def check_partial_files(partial_files_dir: Path):
@@ -63,7 +64,7 @@ def check_partial_files(partial_files_dir: Path):
         actual_number_of_lines = sum(1 for _ in print_file.open())
         if int(batch_quantity) == actual_number_of_lines:
             if not check_partial_has_no_duplicates(print_file, pack_code):
-                logger.error('Quarantining print file with duplicates', partial_file_name=print_file.name)
+                logger.info('Quarantining print file with duplicates', partial_file_name=print_file.name)
                 quarantine_partial_file(print_file)
                 return
             process_complete_file(print_file, pack_code)
@@ -89,30 +90,6 @@ def copy_files_to_sftp(file_paths: Collection[Path], remote_directory):
             sftp_client.put_file(local_path=str(file_path), filename=file_path.name)
         logger.info(f'All {len(file_paths)} files successfully written to SFTP remote',
                     sftp_directory=sftp_client.sftp_directory)
-
-
-def generate_manifest_file(manifest_file_path: Path, print_file_path: Path, productpack_code: str):
-    manifest = create_manifest(print_file_path, productpack_code)
-    manifest_file_path.write_text(json.dumps(manifest))
-
-
-def create_manifest(print_file_path: Path, productpack_code: str) -> dict:
-    return {
-        'schemaVersion': '1',
-        'description': PRODUCTPACK_CODE_TO_DESCRIPTION[productpack_code],
-        'dataset': PACK_CODE_TO_DATASET[productpack_code],
-        'version': '1',
-        'manifestCreated': datetime.utcnow().isoformat(),
-        'sourceName': 'ONS_RM',
-        'files': [
-            {
-                'name': print_file_path.name,
-                'relativePath': './',
-                'sizeBytes': str(print_file_path.stat().st_size),
-                'md5Sum': hashlib.md5(print_file_path.read_text().encode()).hexdigest()
-            }
-        ]
-    }
 
 
 def check_partial_has_no_duplicates(partial_file_path: Path, pack_code: str):
